refactor: ContentType enum으로 수정
