+++ tomcat/src/main/java/org/apache/coyote/http11/httpResponse/StatusLine.java
@@ -1,7 +1,5 @@
 package org.apache.coyote.http11.httpResponse;
 
-import org.apache.coyote.http11.HttpStatus;
-
 public class StatusLine {
 
     private final String protocolVersion;
